Moved deprecation decorator to netket._core; Rephrased brief function descriptions
